Move disco plugin into folder and import lodash utilities separately
