+++ src/headless/headless.js
@@ -9,7 +9,7 @@ import "./plugins/caps.js";         // XEP-0115 Entity Capabilities
 import "./plugins/carbons.js";      // XEP-0280 Message Carbons
 import "./plugins/chat/index.js";   // RFC-6121 Instant messaging
 import "./plugins/chatboxes/index.js";
-import "./plugins/disco.js";        // XEP-0030 Service discovery
+import "./plugins/disco/index.js";  // XEP-0030 Service discovery
 import "./plugins/headlines.js";    // Support for headline messages
 import "./plugins/mam/index.js";    // XEP-0313 Message Archive Management
 import "./plugins/muc/index.js";    // XEP-0045 Multi-user chat
