+++ src/plugins/register/panel.js
@@ -1,4 +1,5 @@
 import log from "@converse/headless/log";
+import pick from "lodash-es/pick";
 import tpl_form_input from "templates/form_input.js";
 import tpl_form_url from "templates/form_url.js";
 import tpl_form_username from "templates/form_username.js";
@@ -8,7 +9,6 @@ import utils from "@converse/headless/utils/form";
 import { ElementView } from "@converse/skeletor/src/element";
 import { __ } from 'i18n';
 import { _converse, api, converse } from "@converse/headless/core";
-import { pick } from "lodash-es";
 import { render } from 'lit-html';
 
 // Strophe methods for building stanzas
